# openapi.yaml
openapi: 3.0.3
info:
  title: Postcredits API
  description: Unified media tracking platform for books and movies with SQL-based recommendations
  version: 1.0.0
  contact:
    name: Team Postcredits
    email: postcredits@tiet.edu.in
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /ratings:
    get:
      tags:
        - Ratings
      summary: Get user's ratings
      operationId: getUserRatings
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [book, movie]
      responses:
        '200':
          description: User's ratings
    post:
      tags:
        - Ratings
      summary: Rate a book or movie
      operationId: createRating
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workId
                - score
              properties:
                workId:
                  type: integer
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '201':
          description: Rating created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
  /ratings/{ratingId}:
    put:
      tags:
        - Ratings
      summary: Update a rating
      operationId: updateRating
      security:
        - BearerAuth: []
      parameters:
        - name: ratingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - score
              properties:
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
      responses:
        '200':
          description: Rating updated
    delete:
      tags:
        - Ratings
      summary: Delete a rating
      operationId: deleteRating
      security:
        - BearerAuth: []
      parameters:
        - name: ratingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Rating deleted

  # ==================== Statistics (Spotistats) ====================
...
  /stats/years:
    get:
      tags:
        - Statistics
      summary: Get user's rating statistics by year
      operationId: getYearStats
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [book, movie]
      responses:
        '200':
          description: Highest rated years
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    year:
                      type: integer
                    avgRating:
                      type: number

  # ==================== Recommendations ====================
  /recommendations/blend:
    get:
      tags:
        - Recommendations
      summary: Get personalized recommendations
      operationId: getBlendRecommendations
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [book, movie, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Blend recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    workId:
                      type: integer
                    title:
                      type: string
                    type:
                      type: string
                    similarity:
                      type: number
  /recommendations/similarity/{userId}:
    get:
      tags:
        - Recommendations
      summary: Calculate similarity with another user
      operationId: getSimilarity
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Similarity score
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  similarityScore:
                    type: number
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Rating:
      type: object
      properties:
        ratingId:
          type: integer
        userId:
          type: integer
        workId:
          type: integer
        workType:
          type: string
        score:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...